requirements.txt
python-chess>=1.10.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
rich>=13.7.0
click>=8.1.0
